Added support for SHA256. Fixed warnings.

+++ src/WiFiServer.cpp
@@ -120,7 +120,7 @@ size_t WiFiServer::write(const uint8_t *buffer, size_t size)
 	for (int sock = 0; sock < TCP_SOCK_MAX; sock++) {
 		client = WiFi._client[sock];
 		if (client && client->_flag & SOCKET_BUFFER_FLAG_CONNECTED) {
-			if (((client->_flag >> SOCKET_BUFFER_FLAG_PARENT_SOCKET_POS) & 0xff) == _socket) {
+			if (((client->_flag >> SOCKET_BUFFER_FLAG_PARENT_SOCKET_POS) & 0xff) == (uint8)_socket) {
 				n += client->write(buffer, size);
 			}
 		}
